pyproject.toml
[project]
name = "cc-slack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "claude-code-sdk>=0.0.25",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "slack-bolt>=1.28.0",
]
